feat: Implement Disciplined Walk-Forward Optimization with Robustness Scoring
- Integrated Walk-Forward Efficiency (WFE) into the robustness score.
- Enhanced parameter stability analysis with scale-robust epsilon perturbations.
- Refined regime consistency logic with a neutral fallback for single-regime data.
- Added comprehensive unit tests for window alignment, WFE, and stability.
- Updated Pydantic models and docstrings to institutional standards.
- Updated documentation in docs/research/WALK_FORWARD_OPTIMIZATION.md.

+++ src/research/hyperopt_walkforward.py
@@ -294,7 +294,7 @@ class WalkForwardOptimizer:
                         new_val = original_val + (direction * delta)
 
                         if is_int:
-                            new_val = int(round(new_val))
+                            new_val = round(new_val)
 
                         # Skip if no actual change
                         if new_val == original_val:
